requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
keras
xgboost
joblib
fpdf2
tabulate
